+++ sdk/src/main/java/ly/count/android/sdk/CountlyConfig.java
@@ -208,6 +208,7 @@ public class CountlyConfig {
 
     // If set to true, immediate requests will use serial AsyncTask executor instead of the thread pool
     boolean useSerialExecutor = false;
+    WebViewDisplayOption webViewDisplayOption = WebViewDisplayOption.IMMERSIVE;
 
     /**
      * THIS VARIABLE SHOULD NOT BE USED
@@ -1052,6 +1053,20 @@ public class CountlyConfig {
             tempRequestTimeoutDuration = 1;
         }
         this.requestTimeoutDuration = tempRequestTimeoutDuration;
+        return this;
+    }
+
+    /**
+     * Set the webview display option for Content and Feedback Widgets
+     *
+     * @param displayOption IMMERSIVE for full screen with hidden system UI, or
+     *                      SAFE_AREA to use app usable area and not overlap system UI
+     * @return config content to chain calls
+     */
+    public synchronized CountlyConfig setWebviewDisplayOption(WebViewDisplayOption displayOption) {
+        if (displayOption != null) {
+            this.webViewDisplayOption = displayOption;
+        }
         return this;
     }